Add files via upload
digit10.txt is the file which contains only 0-9 each digits in each line and which is read by the script and pushed into the bash array and the corresponding random sequence of 10 decimal digits are been generated. I made wrong loop count as 3225600 in script which is supposed to be 3265920 = 9 * 9!. Please correct the loop count as 3265920 before you run it.